create InvocationMetricCard
